refactored table component
