+++ ReactComponents/src/components/StyledFontAwesome.tsx
@@ -1,0 +1,6 @@
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import styled from "styled-components";
+
+export const StyledFontAwesome = styled(FontAwesomeIcon)`
+  cursor: pointer;
+`;
