+++ ReactComponents/src/views/Table/TableMetaData.tsx
@@ -1,58 +1,14 @@
-import {
-  faCaretUp,
-  faCaretDown,
-  faSearch,
-  faFilter,
-} from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { columnConfig, SortOrder } from "./types";
+import { columnConfig } from "./types";
 import { Wrapper } from "components/Wrapper";
-import { useCallback, useContext, useEffect, useState } from "react";
-import SearchInBox from "components/SearchInBox";
+import { useContext } from "react";
 import { TableContext } from "./contexts/TableContext";
-import FilterBox from "components/FilterBox";
+import FilterBox from "./FilterBox";
+import SearchInBox from "./SearchInBox";
+import Sorter from "./Sorter";
 
 function TableHeader<T>({ column }: { column: columnConfig<T> }) {
-  const { handleSorting, loading } = useContext(TableContext);
-  const [sortOrder, setSortOrder] = useState<SortOrder | undefined>(
-    column.defaultSortOrder
-  );
-  const [isSearchOpen, setIsSearchOpen] = useState<boolean>(false);
-  const [isOptionsOpen, setIsOptionsOpen] = useState<boolean>(false);
-
-  useEffect(() => {
-    if ((column.defaultSortOrder || sortOrder) && !loading) {
-      handleSorting(column.dataIndex, sortOrder);
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [sortOrder, loading]);
-
-  const handleClick = () => {
-    if (column.sorter) {
-      setSortOrder((prev) => {
-        if (prev === SortOrder.ascending) {
-          return SortOrder.descending;
-        }
-        if (prev === SortOrder.descending) {
-          return undefined;
-        }
-
-        return SortOrder.ascending;
-      });
-    }
-  };
-
-  const handleSearchOpen = () => {
-    setIsSearchOpen((prev) => !prev);
-  };
-
-  const handleFiltersOpen = useCallback(() => {
-    setIsOptionsOpen((prev) => !prev);
-  }, []);
-
   return (
-    <th onClick={handleClick}>
+    <th>
       <Wrapper
         display="flex"
         gap="10px"
@@ -61,65 +17,14 @@ function TableHeader<T>({ column }: { column: columnConfig<T> }) {
       >
         <div>{column.title}</div>
         <div>
-          <div style={{ display: "flex", gap: "10px" }}>
-            {column.sorter && (
-              <div
-                style={{
-                  position: "relative",
-                }}
-              >
-                <div
-                  style={{
-                    position: "absolute",
-                    top: "-11px",
-                    left: "1px",
-                    color:
-                      sortOrder === SortOrder.ascending ? "blue" : "inherit",
-                  }}
-                >
-                  <FontAwesomeIcon icon={faCaretUp} />
-                </div>
-                <div
-                  style={{
-                    position: "absolute",
-                    top: "-5px",
-                    left: "1px",
-                    color:
-                      sortOrder === SortOrder.descending ? "blue" : "inherit",
-                  }}
-                >
-                  <FontAwesomeIcon icon={faCaretDown} />
-                </div>
-              </div>
-            )}
+          {column.sorter && <Sorter<T> column={column} />}
+          <div>
+            {column.searchInColumn && <SearchInBox column={column.dataIndex} />}
           </div>
           <div>
-            {column.searcInColumn && (
-              <div>
-                <FontAwesomeIcon icon={faSearch} onClick={handleSearchOpen} />
-              </div>
-            )}
-          </div>
-          <div>
-            {column.dynamicFilters && (
-              <div>
-                <FontAwesomeIcon icon={faFilter} onClick={handleFiltersOpen} />
-              </div>
-            )}
+            {column.dynamicFilters && <FilterBox column={column.dataIndex} />}
           </div>
         </div>
-        {column.searcInColumn && isSearchOpen && (
-          <SearchInBox
-            handleSearchOpen={handleSearchOpen}
-            column={column.dataIndex}
-          />
-        )}
-        {column.dynamicFilters && isOptionsOpen && (
-          <FilterBox
-            column={column.dataIndex}
-            handleFiltersOpen={handleFiltersOpen}
-          />
-        )}
       </Wrapper>
     </th>
   );
@@ -127,12 +32,6 @@ function TableHeader<T>({ column }: { column: columnConfig<T> }) {
 
 export default function TableMetaData() {
   const { columnConfig } = useContext(TableContext);
-
-  useEffect(() => {
-    return () => {
-      localStorage.setItem("checked", "");
-    };
-  }, []);
 
   return (
     <tr>
