+++ ReactComponents/src/views/Table/table.css
@@ -18,7 +18,6 @@ td {
 th {
   border-right: 1px solid #f5f5f5;
   line-height: 10px;
-  cursor: pointer;
 }
 
 tbody > tr > td {
